Add test that nile node calls starknet-devnet with the correct args
This test aims to cover the recently added `nile node` --seed option.
This argument allows callers to get the same prefunded accounts on
successive calls to `nile node`.

Testing this effect directly would require more code than it's worth, as
it the heavy lifting is implemented by `starknet-devnet`, not `nile`. So
we choose to follow the `test_debug` precedent of confirming
`subprocess` cli calls are made correctly.

These checks are made in a new test case, rather than using `test_node`,
because `unittest.mocks.patch` does not work as expected within a
`multiproccessing.Process`, which is used in `test_node`.

`test_node` is renamed to add specificity, and --seed test case was added
to check that `--seed` is indeed a valid `starknet-devnet` option.